changed rating value to double, and started creating jwt service and filter to authenticate requests

+++ src/main/java/org/backend/gamelibwebapp/dto/RatingRequest.java
@@ -1,4 +1,4 @@
 package org.backend.gamelibwebapp.dto;
 
-public record RatingRequest(int value,Long gameId, Long userId) {
+public record RatingRequest(double value,Long gameId, Long userId) {
 }
